Remove top-selling collections from project scope
Since collections are not atomic, the database structure to store a list of top-selling collections (as well as tying them to products & orders) would over-complicate the project. As such, it has been removed from the project's scope at this time. Code referring to top Collections has either been commented out or deleted.